+++ Examples/PSPDFCatalog/Subclassing/PSCTimingTestViewController.m
@@ -46,7 +46,7 @@
 - (void)showRandomPDFController {
     if (self.documents.count == 0) return;
 
-    NSUInteger index = arc4random_uniform(self.documents.count);
+    NSUInteger index = arc4random_uniform((u_int32_t)self.documents.count);
     PSPDFDocument *randomDocument = self.documents[index]; // create a copy to better simulate memory issues
     NSLog(@"Loading %@", randomDocument);
